refactor: modularize application logic by migrating services and components to separate files
