--- recipes.js
@@ -1,295 +1,0 @@
-export const recipes = {
-    recipe1: {
-        title: 'Blueberry Pancakes',
-        image: 'https://images.unsplash.com/photo-1567620905732-2d1ec7ab7445?auto=format&fit=crop&w=800&q=80',
-        description: 'Fluffy pancakes loaded with fresh blueberries and served with maple syrup.',
-        ingredients: [
-            '1 cup all-purpose flour',
-            '2 tbsp sugar',
-            '1 tsp baking powder',
-            '1/2 tsp baking soda',
-            '1 cup buttermilk',
-            '1 large egg',
-            '1 cup fresh blueberries'
-        ],
-        instructions: '1. Whisk dry ingredients in a bowl.\n2. Whisk wet ingredients in another bowl.\n3. Combine wet and dry mixtures; fold in blueberries.\n4. Cook on a hot buttered griddle until golden brown.',
-        macros: { calories: 350, protein: 8, fat: 10, carbs: 55 },
-        category: 'breakfast',
-        servings: 2,
-        totalTime: '20 min',
-        difficulty: 'Easy'
-    },
-    recipe2: {
-        title: 'Grilled Salmon Salad',
-        image: 'https://images.unsplash.com/photo-1546069901-ba9599a7e63c?auto=format&fit=crop&w=800&q=80',
-        description: 'A healthy salad featuring grilled salmon, avocado, and mixed greens.',
-        ingredients: [
-            '2 salmon fillets',
-            '5 oz mixed greens',
-            '1 avocado, sliced',
-            '1/2 cucumber, sliced',
-            '1/4 cup cherry tomatoes',
-            '2 tbsp olive oil',
-            '1 tbsp lemon juice'
-        ],
-        instructions: '1. Season salmon with salt and pepper.\n2. Grill salmon for 4-5 minutes per side.\n3. Toss greens and veggies with oil and lemon juice.\n4. Top salad with grilled salmon.',
-        macros: { calories: 450, protein: 40, fat: 25, carbs: 10 },
-        category: 'lunch',
-        servings: 2,
-        totalTime: '25 min',
-        difficulty: 'Medium'
-    },
-    recipe3: {
-        title: 'Chocolate Lava Cake',
-        image: 'https://images.unsplash.com/photo-1624353365286-3f8d62daad51?auto=format&fit=crop&w=800&q=80',
-        description: 'Decadent chocolate cake with a gooey molten center.',
-        ingredients: [
-            '4 oz semi-sweet chocolate',
-            '1/2 cup butter',
-            '1 cup powdered sugar',
-            '2 eggs + 2 yolks',
-            '6 tbsp flour',
-            'Cocoa powder for dusting'
-        ],
-        instructions: '1. Melt chocolate and butter together.\n2. Whisk eggs and sugar until pale.\n3. Fold chocolate into egg mixture, then fold in flour.\n4. Pour into greased ramekins and bake at 425°F (220°C) for 13 mins.',
-        macros: { calories: 520, protein: 7, fat: 34, carbs: 50 },
-        category: 'dessert',
-        servings: 4,
-        totalTime: '30 min',
-        difficulty: 'Medium'
-    },
-    recipe4: {
-        title: 'Creamy Pesto Pasta',
-        image: 'https://images.unsplash.com/photo-1473093295043-cdd812d0e601?auto=format&fit=crop&w=800&q=80',
-        description: 'Al dente pasta tossed in a rich basil pesto sauce topped with parmesan.',
-        ingredients: [
-            '1 lb pasta (penne or fusilli)',
-            '1 cup fresh basil pesto',
-            '1/2 cup heavy cream',
-            '1/2 cup grated Parmesan cheese',
-            '1/4 cup pine nuts, toasted',
-            'Salt and pepper to taste'
-        ],
-        instructions: '1. Cook pasta according to package directions.\n2. Simmer pesto and cream in a saucepan for 2 mins.\n3. Toss pasta with the sauce.\n4. Garnish with parmesan and pine nuts.',
-        macros: { calories: 600, protein: 18, fat: 35, carbs: 70 },
-        category: 'dinner',
-        servings: 4,
-        totalTime: '30 min',
-        difficulty: 'Easy'
-    },
-    recipe5: {
-        title: 'Avocado Toast',
-        image: 'https://images.unsplash.com/photo-1541519227354-08fa5d50c44d?auto=format&fit=crop&w=800&q=80',
-        description: 'Crispy sourdough topped with smashed avocado, radish, and chili flakes.',
-        ingredients: [
-            '2 slices sourdough bread',
-            '1 ripe avocado',
-            '1 tsp lemon juice',
-            '2 radishes, thinly sliced',
-            'Red pepper flakes',
-            'Sea salt and black pepper'
-        ],
-        instructions: '1. Toast the bread until golden.\n2. Mash avocado with lemon juice, salt, and pepper.\n3. Spread onto toast.\n4. Top with radish slices and chili flakes.',
-        macros: { calories: 320, protein: 9, fat: 18, carbs: 35 },
-        category: 'breakfast',
-        servings: 2,
-        totalTime: '10 min',
-        difficulty: 'Easy'
-    },
-    recipe7: {
-        title: 'Spaghetti Carbonara',
-        image: 'https://images.unsplash.com/photo-1612874742237-6526221588e3?auto=format&fit=crop&w=800&q=80',
-        description: 'Classic Italian pasta dish with eggs, cheese, pancetta, and pepper.',
-        ingredients: [
-            '12 oz spaghetti',
-            '4 oz pancetta or guanciale, diced',
-            '2 large eggs',
-            '1/2 cup grated Pecorino Romano',
-            '1/2 cup grated Parmesan',
-            'Freshly ground black pepper'
-        ],
-        instructions: '1. Cook pasta in salted water until al dente.\n2. Crispen pancetta in a pan.\n3. Whisk eggs and cheese in a bowl.\n4. Toss hot pasta with pancetta, then remove from heat and mix in egg mixture quickly to create a creamy sauce.',
-        macros: { calories: 400, protein: 15, fat: 20, carbs: 45 },
-        category: 'dinner',
-        servings: 4,
-        totalTime: '25 min',
-        difficulty: 'Medium'
-    },
-    recipe8: {
-        title: 'Classic Caesar Salad',
-        image: 'https://images.unsplash.com/photo-1550304943-4f24f54ddde9?auto=format&fit=crop&w=800&q=80',
-        description: 'Crisp romaine lettuce with creamy dressing, croutons, and parmesan.',
-        ingredients: [
-            '1 head romaine lettuce, chopped',
-            '1/2 cup croutons',
-            '1/4 cup grated Parmesan cheese',
-            '1/4 cup Caesar dressing',
-            '1 lemon wedge',
-            'Anchovy fillets (optional)'
-        ],
-        instructions: '1. Wash and dry lettuce thoroughly.\n2. Toss lettuce with dressing until coated.\n3. Add croutons and parmesan cheese.\n4. Serve immediately with a squeeze of lemon.',
-        macros: { calories: 300, protein: 8, fat: 22, carbs: 12 },
-        category: 'lunch',
-        servings: 2,
-        totalTime: '15 min',
-        difficulty: 'Easy'
-    },
-    recipe9: {
-        title: 'Vegetable Stir Fry',
-        image: 'https://images.unsplash.com/photo-1512058564366-18510be2db19?auto=format&fit=crop&w=800&q=80',
-        description: 'Colorful mix of fresh vegetables wok-tossed in a savory soy-ginger sauce.',
-        ingredients: [
-            '2 cups mixed vegetables (broccoli, peppers, snap peas)',
-            '1 tbsp vegetable oil',
-            '2 cloves garlic, minced',
-            '1 tbsp ginger, grated',
-            '2 tbsp soy sauce',
-            '1 tsp sesame oil',
-            'Rice for serving'
-        ],
-        instructions: '1. Heat oil in a wok or large pan.\n2. Stir-fry garlic and ginger for 30 seconds.\n3. Add vegetables and stir-fry for 5-7 minutes until tender-crisp.\n4. Stir in soy sauce and sesame oil; serve over rice.',
-        macros: { calories: 250, protein: 6, fat: 10, carbs: 35 },
-        category: 'dinner',
-        servings: 3,
-        totalTime: '20 min',
-        difficulty: 'Easy'
-    },
-    recipe10: {
-        title: 'Mushroom Risotto',
-        image: 'https://images.unsplash.com/photo-1476124369491-e7addf5db371?auto=format&fit=crop&w=800&q=80',
-        description: 'Creamy Italian rice dish cooked with white wine, parmesan, and mushrooms.',
-        ingredients: [
-            '1 cup Arborio rice',
-            '4 cups vegetable broth, warm',
-            '8 oz mushrooms, sliced',
-            '1/2 cup white wine',
-            '1/2 onion, diced',
-            '2 tbsp butter',
-            '1/4 cup Parmesan cheese'
-        ],
-        instructions: '1. Sauté onion and mushrooms in butter until soft.\n2. Add rice and toast for 1 minute.\n3. Add wine and cook until absorbed.\n4. Gradually add warm broth, one ladle at a time, stirring constantly until rice is creamy and tender. Stir in cheese.',
-        macros: { calories: 420, protein: 10, fat: 14, carbs: 58 },
-        category: 'dinner',
-        servings: 4,
-        totalTime: '40 min',
-        difficulty: 'Hard'
-    },
-    recipe11: {
-        title: 'Greek Yogurt Parfait',
-        image: 'https://images.unsplash.com/photo-1488477181946-6428a0291777?auto=format&fit=crop&w=800&q=80',
-        description: 'Layers of creamy Greek yogurt, fresh fruits, and crunchy granola.',
-        ingredients: [
-            '1 cup Greek yogurt',
-            '1/2 cup granola',
-            '1/2 cup fresh berries (strawberries, blueberries)',
-            '1 tbsp honey',
-            '1 tsp chia seeds',
-            'Mint leaves for garnish'
-        ],
-        instructions: '1. Place half of the yogurt in a glass.\n2. Add a layer of granola and fruit.\n3. Repeat layers.\n4. Drizzle with honey and top with chia seeds.',
-        macros: { calories: 300, protein: 15, fat: 8, carbs: 40 },
-        category: 'breakfast',
-        servings: 2,
-        totalTime: '10 min',
-        difficulty: 'Easy'
-    },
-    recipe12: {
-        title: 'Homemade Beef Tacos',
-        image: 'https://images.unsplash.com/photo-1551504734-5ee1c4a1479b?auto=format&fit=crop&w=800&q=80',
-        description: 'Flavorful ground beef served in soft tortillas with fresh toppings.',
-        ingredients: [
-            '1 lb ground beef',
-            '1 packet taco seasoning',
-            '8 small corn tortillas',
-            '1/2 cup diced onion',
-            '1/2 cup cilantro, chopped',
-            'Lime wedges',
-            'Salsa'
-        ],
-        instructions: '1. Brown ground beef in a skillet; drain fat.\n2. Add taco seasoning and water according to packet instructions; simmer.\n3. Warm tortillas.\n4. Assemble tacos with beef and toppings.',
-        macros: { calories: 500, protein: 25, fat: 22, carbs: 45 },
-        category: 'dinner',
-        servings: 4,
-        totalTime: '30 min',
-        difficulty: 'Easy'
-    },
-    recipe13: {
-        title: 'Caprese Salad',
-        image: 'https://images.unsplash.com/photo-1592417817098-8fd3d9eb14a5?auto=format&fit=crop&w=800&q=80',
-        description: 'Simple and refreshing salad with fresh mozzarella, tomatoes, and basil.',
-        ingredients: [
-            '2 large tomatoes, sliced',
-            '8 oz fresh mozzarella, sliced',
-            '1/2 cup fresh basil leaves',
-            '2 tbsp olive oil',
-            '1 tbsp balsamic glaze',
-            'Salt and pepper'
-        ],
-        instructions: '1. Arrange tomato and mozzarella slices on a platter.\n2. Tuck basil leaves between slices.\n3. Drizzle with olive oil and balsamic glaze.\n4. Season with salt and pepper.',
-        macros: { calories: 280, protein: 12, fat: 20, carbs: 8 },
-        category: 'lunch',
-        servings: 2,
-        totalTime: '15 min',
-        difficulty: 'Easy'
-    },
-    recipe14: {
-        title: 'Apple Pie',
-        image: 'https://images.unsplash.com/photo-1568571780765-9276ac8b75a2?auto=format&fit=crop&w=800&q=80',
-        description: 'Traditional dessert filling your home with the scent of cinnamon and baked apples.',
-        ingredients: [
-            '1 double pie crust',
-            '6 cups apples, peeled and sliced',
-            '3/4 cup sugar',
-            '1 tsp cinnamon',
-            '1 tbsp butter',
-            '1 egg (for wash)'
-        ],
-        instructions: '1. Toss apples with sugar and cinnamon.\n2. Place bottom crust in pie dish and fill with apples. Dot with butter.\n3. Top with second crust, seal edges, and cut vents.\n4. Brush with egg wash and bake at 375°F (190°C) for 50 mins.',
-        macros: { calories: 350, protein: 3, fat: 15, carbs: 55 },
-        category: 'dessert',
-        servings: 8,
-        totalTime: '80 min',
-        difficulty: 'Medium'
-    },
-    recipe15: {
-        title: 'Banana Bread',
-        image: 'https://images.unsplash.com/photo-1603532648955-039310d9ed75?auto=format&fit=crop&w=800&q=80',
-        description: 'Moist and sweet loaf cake made from ripe mashed bananas.',
-        ingredients: [
-            '3 ripe bananas, mashed',
-            '1/3 cup melted butter',
-            '3/4 cup sugar',
-            '1 egg, beaten',
-            '1 tsp vanilla extract',
-            '1 tsp baking soda',
-            '1 1/2 cups all-purpose flour'
-        ],
-        instructions: '1. Mix mashed bananas with butter.\n2. Stir in sugar, egg, and vanilla.\n3. Sprinkle baking soda and salt over mixture and mix.\n4. Stir in flour. Pour into loaf pan and bake at 350°F (175°C) for 60 mins.',
-        macros: { calories: 220, protein: 4, fat: 8, carbs: 35 },
-        category: 'breakfast',
-        servings: 8,
-        totalTime: '75 min',
-        difficulty: 'Easy'
-    },
-    recipe16: {
-        title: 'Classic French Toast',
-        image: 'https://images.unsplash.com/photo-1484723091739-30a097e8f929?auto=format&fit=crop&w=800&q=80',
-        description: 'Bread slices soaked in egg and milk mixture, fried to golden perfection.',
-        ingredients: [
-            '4 slices thick bread',
-            '2 eggs',
-            '1/2 cup milk',
-            '1 tsp vanilla extract',
-            '1/2 tsp cinnamon',
-            'Butter for frying'
-        ],
-        instructions: '1. Whisk eggs, milk, vanilla, and cinnamon in a shallow dish.\n2. Dip bread slices into mixture, coating both sides.\n3. Cook on a buttered skillet over medium heat until golden brown on both sides.\n4. Serve with syrup and fruit.',
-        macros: { calories: 380, protein: 12, fat: 14, carbs: 45 },
-        category: 'breakfast',
-        servings: 2,
-        totalTime: '20 min',
-        difficulty: 'Easy'
-    }
-};
-
-console.log('Recipes loaded:', Object.keys(recipes).length);
